[refactor] #22 ElectionUtil 삭제, PoliticianInfoService로 코드 이동
